Setting up style for background, titles and text
Most of the layout has now the GWDG style.
There are still some details to change, but they are not important.
Some @import paths were also corrected (missing modifications when the full default theme was imported and activated).

Timelines have an independent stylesheet not part of the default theme; to include it in the GWDG theme will involve big effort.